requirements.txt
# This file may be used to create an environment using:
opencensus-ext-azure>=1.0.7
databricks-connect==10.4.0b0
databricks-api==0.8.0
typeguard>=2.12.0
pytest>=6.2.3
jupyter>=1.0.0
python-dotenv>=0.17.0
pypandoc>=1.4
pdoc3>=0.7.4
pandas>=1.2.4
setuptools>=56.0.0
pydataset>=0.2.0
scikit-learn>=0.24.1
PyArrow>=0.15.1
databricks-cli==0.17.0
azure-cli==2.40.0
